requirements.txt
Django==5.2.8
ipython==8.24.0
djangorestframework==3.16.1
Markdown==3.6
django-filter==24.2
django-environ==0.11.2
mysqlclient==2.2.7
django-cors-headers==4.9.0
gunicorn==23.0.0
pillow==12.0.0
psycopg==3.2.12
requests==2.32.5
boto3==1.40.74
django-storages==1.14.6
